Refactor: Move tree_clvsa_wrapper to utils/ml_common/models
